Update with MainStreetHub needs.
This update contains two principle parts.  First is the upgrade to the
newest version of the AWS Java SDK.  This enables us to use a better
credentials provider chain that will read a ~/.aws/credentials file
if it's present.

Secondly, we update the PutObject API calls to also grant the bucket
owner full access to any objects that were uploaded.  This is
important because, at MainStreetHub at least, our build system has
credentials to a different AWS account from the one that owns the
S3 bucket that we're writing artifacts into.  If we don't do this
then we end up in a situation where we can write artifacts just
fine to the bucket, but can't read them!